+++ components/posts/list/item.tsx
@@ -1,4 +1,4 @@
-import { Post } from "contentlayer/generated";
+import { Post } from "@velite";
 import moment from "moment";
 import Link from "next/link";
 import { format } from "date-fns";