+++ lib/remark-next-images.ts
@@ -1,4 +1,3 @@
-import { RawDocumentData } from "contentlayer/source-files";
 import fs from "fs/promises";
 import { Element } from "hast";
 import crypto from "node:crypto";
@@ -7,10 +6,6 @@ import sharp from "sharp";
 import { visit } from "unist-util-visit";
 import { VFile } from "vfile";
 import createPlaceholder from "./data-blur";
-
-type FileData = {
-  rawDocumentData: RawDocumentData;
-};
 
 const fileChecksum = async (file: string) => {
   try {
@@ -26,16 +21,6 @@ const checksum = (content: Buffer) => {
 
 const findRoot = (file: VFile) => {
   return file.dirname || process.cwd();
-};
-
-const findPath = (file: VFile, image: Element) => {
-  if (file.dirname) {
-    return path.join(
-      (file.data.rawDocumentData as any).flattenedPath.replace(/posts\//, ""),
-      image.properties?.src as string
-    );
-  }
-  return (image.properties?.src as string) || "";
 };
 
 const copy = async (source: string, sha256sum: string, target: string) => {
@@ -58,7 +43,6 @@ const metadata = async (source: string, pathname: string) => {
   }
 
   const src = "/" + pathname;
-
   const sha256 = checksum(content);
 
   const blurDataURL = await createPlaceholder(image);
@@ -76,11 +60,12 @@ const metadata = async (source: string, pathname: string) => {
 const processImage = async (
   options: Options,
   file: VFile,
-  node: Element
+  node: Element,
 ): Promise<void> => {
   const root = findRoot(file);
-  const pathname = findPath(file, node);
+  const pathname = (node.properties?.src as string) || "";
   const source = path.join(root, pathname);
+
   const meta = await metadata(source, pathname);
   if (!meta) {
     return;
@@ -92,6 +77,11 @@ const processImage = async (
   if (!node.properties) {
     node.properties = {};
   }
+
+  // console.debug(
+  //   `${JSON.stringify(node.properties)} ${JSON.stringify(meta.props)}`,
+  // );
+
   node.properties = {
     ...node.properties,
     ...meta.props,
@@ -102,16 +92,21 @@ type Options = {
   publicDir: string;
 };
 
-const staticImages = (options: Options) => (tree: any, file: VFile, done: () => any) => {
-  const tasks: Promise<void>[] = [];
+/**
+ * Remark plugin that adds image metadata to a img node
+ * for Next.js optimization.
+ */
+const staticImages =
+  (options: Options) => (tree: any, file: VFile, done: () => any) => {
+    const tasks: Promise<void>[] = [];
 
-  visit(tree, "element", (node) => {
-    if (node.tagName === "img") {
-      tasks.push(processImage(options, file, node));
-    }
-  });
+    visit(tree, "element", (node) => {
+      if (node.tagName === "img") {
+        tasks.push(processImage(options, file, node));
+      }
+    });
 
-  Promise.all(tasks).then(() => done());
-};
+    Promise.all(tasks).then(() => done());
+  };
 
 export default staticImages;